Improve input text handling
Add dummy call to check if input is Caps Lock to prevent "null" from being entered. There is probably a better way to handle this.
Check if the character is null first, then check if empty or not, and check if the character is a letter, space, number, symbol, or otherwise, before entering it into the input.