Middle mouse on Linux crash's zui.Nodes
This fixes middle mouse button being considered a paste event; i.e. we crash when we have zui.Nodes open and we middle mouse click to pan since it thinks we want to paste on Linux(this can vary on flavors of linux but it is a default on some flavors). 

This could crash if what we have in the pastebin isn't json and we arrive at this [line](https://github.com/armory3d/zui/blob/master/Sources/zui/Nodes.hx#L389)