Reset window scroll when closing the popup
Right now if the popup needed to scroll down to say go to a apply or cancel button, next time we would open a popup the window would be scrolled to as far as was set beforehand. This addresses the issue by reset6ing the scroll offset when closing the Popup.